Set up Git-LFS (2GB storage limit)
Large File Storage - better handling of images, audio etc in regard to
space usage. Currently looks for PNG, SVG, and MP3 files - if you add
a massive attribute that's something else, talk to me and we'll probably
add it to this list.

If this gets too full, we can custom pick the files to stop tracking with
it.